fix: Add Minimum-free ZRAM configuration
It assures that this value will get properly configured for systems with ≥6.75GBs of RAM.

As this article shows, too low of a value is not good, and too high of a value is not good.

https://linuxhint.com/vm_min_free_kbytes_sysctl/

It is the only remaining non-applied ZRAM tweak from Pop-OS, which they included in their zram service.

pop-os/default-settings@7fcad35#diff-964fc7d24025670fb776c01ef2943c84e637a3675c164696aebae19d86c8d46eR62

This checks if minimum-free ZRAM value is 1% of the maximum RAM of the current system, hence why this must go to the hardware service compared to simple sysctl config change.