fix(user-setup): Improve language-sensible templates setup (#1180)
* fix(user-setup): Use correct templates directory depending of language

Use 'xdg-user-dirs' to obtain a localized path to templates directory (with '$HOME/Templates' as fallback.
Should fix any related issues with Gnome nautilus not detecting templates using other language that english.

* fix(bazzite-user-setup): Add checks before creating templates symlinks

Check if the template files exists in `/etc` before creating the symlinks, otherwise do not create them.

Using only '[[ ! -f file ]]' always returns true in the case of a broken symlink, as per quote of `man test`:

> Except  for  -h  and  -L,  all FILE-related tests dereference symbolic links.

In consequence, it checks the symlink file is pointing to, not tne existence of the symlink itself, thus ends up attempting to overwrite the later.

* chore: Increase setup script version

---------

Co-authored-by: Kyle Gospodnetich <me@kylegospodneti.ch>